pyproject.toml
[tool.poetry]
name = "home-3"
version = "0.1.0"
description = ""
authors = ["Roman Desyatskii <roman.desyatskii30@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
poetry = "1.6.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
aerich = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
